Extract radio stream creation code
