+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/streamwrappers/radio/RadioTranscription.scala
@@ -1,0 +1,6 @@
+package com.microsoft.partnercatalyst.fortis.spark.streamwrappers.radio
+
+case class RadioTranscription(
+  text: String,
+  language: String,
+  radioUrl: String)
